Fixed binary read/write for the save files. Switched floats to doubles as the floats were unnecessary. Added proper however rudimentiary handling of entities in the save file and game data. Isolated controls behind the entity defintions. Need to fix controls next. They are not responding.

+++ RaycastEngine/Float2D.h
@@ -2,7 +2,7 @@
 
 struct Float2D
 {
-	float X, Y;
+	double X, Y;
 };
 
 struct Float2D fAdd(struct Float2D, struct Float2D);
